Fixed missing import & added plt.show() calls.
Existing hawkes.py refers to a 'plt' variable, but this must
first be imported to be used. Also, for the plots to display
two separate plt.show() calls are added

Note: the first plots must be closed before the last is shown,
otherwise the final QQ plot is drawn on top of the previous plot).